requirements.txt
streamlit
openai
python-docx
PyPDF2
beautifulsoup4
requests